fix: Dao interface로 변경
