+++ src/main/java/gift/DataInit.java
@@ -1,7 +1,6 @@
 package gift;
 
-import gift.model.Product;
-import gift.model.ProductDao;
+import gift.model.ProductDaoImpl;
 import gift.model.ProductRequest;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,12 +10,12 @@ import org.springframework.stereotype.Component;
 public class DataInit {
 
     @Autowired
-    private ProductDao productDao;
+    private ProductDaoImpl productDaoImpl;
 
     @PostConstruct
     public void init() {
-        productDao.save(new ProductRequest("product1", 10000, "image1.png"));
-        productDao.save(new ProductRequest("product2", 20000, "image2.png"));
-        productDao.save(new ProductRequest("product3", 30000, "image3.png"));
+        productDaoImpl.save(new ProductRequest("product1", 10000, "image1.png"));
+        productDaoImpl.save(new ProductRequest("product2", 20000, "image2.png"));
+        productDaoImpl.save(new ProductRequest("product3", 30000, "image3.png"));
     }
 }
