+++ src/main/java/gift/model/ProductDao.java
@@ -1,48 +1,16 @@
 package gift.model;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import org.springframework.stereotype.Component;
 
-@Component
-public class ProductDao {
+public interface ProductDao {
 
-    private static final Map<Long, Product> store = new HashMap<>();
-    private static long sequence = 0L;
+    Product save(ProductRequest productRequest);
 
-    public Product save(ProductRequest productRequest) {
-        Product product = productRequest.toEntity(++sequence);
-        store.put(sequence, product);
-        return product;
-    }
+    Product findById(Long id);
 
-    public Product findById(Long id) {
-        if (!store.containsKey(id)) {
-            throw new IllegalArgumentException("존재하지 않는 상품입니다.");
-        }
-        return store.get(id);
-    }
+    List<Product> findAll();
 
-    public List<Product> findAll() {
-        return new ArrayList<>(store.values());
-    }
+    Product update(Long id, ProductRequest productRequest);
 
-    public Product update(Long id, ProductRequest productRequest) {
-        if (!store.containsKey(id)) {
-            throw new IllegalArgumentException("존재하지 않는 상품입니다.");
-        }
-        Product product = store.get(id);
-        product.updateProduct(productRequest);
-
-        return product;
-    }
-
-    public void delete(Long id) {
-        if (!store.containsKey(id)) {
-            throw new IllegalArgumentException("존재하지 않는 상품입니다.");
-        }
-        store.remove(id);
-    }
+    void delete(Long id);
 }
